adapter: remove mz_cluster_replicas.status field
This field showed the pod-level replica status, which could be
misleading (status showed "healthy" even if the replica was unable to
complete initialization). We therefore remove that field to avoid
confusion. The per-process replica status is still available in
`mz_internal.mz_cluster_replica_statuses`.

With the removal of the status field, the `mz_cluster_replicas` view
looks the same as the `mz_internal.mz_cluster_replicas_base` table, so
we can remove the latter and make the former a proper system table
again.