[sinks] Add pending interface to storage exports (#15233)
Compaction of storage collections is partially held back by the storage controller's list of exported collections. The coordinator / adapter provides some guidance, on top of which the storage controller layers its own logic based on the output progress of sinks. This is the logic which keeps sinks / tables / materialized views held back far enough so that the sink may write everything out.

However, before this PR, we didn't actually install this dependency until the export (sink) was created in the storage layer. This behavior is incorrect in at least two situations:

CREATE SINK command comes in. We determine an acceptable AS OF and then we instruct the storage layer to create the export, installing the dependency. This dependency should be made known to the storage controller before selecting the AS OF -- it the storage controller can ensure the AS OF stays valid until the usual sink logic takes over.
bootstraping. [sinks] async retry creating sink connection on startup #15228 adds logic to retry sink connections on startup. However, while this retrying occurs, the storage controller does not know about the dependency and can allow the source to be compacted. In this case, when the sink does finally start up, the AS OF is invalid and we enter a crash loop. We avoid that by telling the storage controller immediately -- during the bootstrap process when no frontiers are adjusted.
We use an opaque CreateExportToken to ensure that the pending_export function is called before create_export. It replaces the id argument to create_export. The only other field contained is the from_id which needs to be passed around the coordinator anyway so that we could properly cancel the pending export on error.

I'm not convinced that this is the ideal way to solve the issue (perhaps the logic should live elsewhere?) but I am convinced this is correct -- and so it feels less risky to extend the current behavior than move it around altogether.

Motivation
Necessary to unblock #15228 and re-merging #15146 -- which each close one of the last two issues in the kafka sink epic #11503