storage: make kafka sink more aggressive about crashing on errors
Since PR #14965, our approach to error handling in the Kafka sink is to
crash on unrecoverable errors, letting Kubernetes restart the sink from
scratch. This is not the most efficient approach, as restarting a sink
from scratch is a bit slow, but it is correct and simple to code.

Unfortunately there were still many vestiges of the old approach to
error handling, where, on an unrecoverable error, the sink would log
message to stderr and then drop itself.

This commit eliminates all those vestigial paths. Errors are either
retried or turned into panics as close as possible to where they are
produced.

To enforce that future code follows the new crash-oriented approach to
error handling, the shutdown_flag bool, which causes the sink to drop
itself, is removed from KafkaSinkState, so that flag can only be raised
by the KafkaSinkToken when the user has dropped the sink, rather than by
any method on KafkaSinkState that encounters an unrecoverable error.

Note that the cleanups here are intentionally prioritizing short term
clarity over the long term goal state of this code. In the long term, we
*do* want some method to bubble up unrecoverable errors to the top level
of the sink implementation, where the sink can be rebuilt without
rebooting the whole process. However, for the moment, I found the
multiple layers of retries and stacked shutdown bools in
retry_on_txn_error too hard to reason about. Ideally, when we add the
graceful retry logic , we can find some way to structure the code that's
easier to follow; perhaps the new timely async bridge will help.

Note that this commit also uses a backoff clamp of 30s throughout all
retry loops in the Kafka sink. The old code used 60s in one place and
10m in another; both feel a bit too long to me.

Fix #14098.