storage: refactor load generator batches
Previously transactions in load generators were limited to a single
output and had to produce all rows up front. Refactor so that batches
(transactions) can cross outputs and can generate rows over time. This
is needed for the upcoming TPCH generator, which produces all rows in a
single transaction.

Simplify the output count handling to always use the views length of
load generators, as that correctly indicates whether or not it has
subsources and there's no need to care about individual generators
there.