replicated: track frontiers of arranged logs
This commit fixes an oversight in the `ActiveReplication` client, where
frontiers of persisted logs would be correctly propagted but frontiers
of arranged logs would be ignored.

Arranged logs are shared between all replicas (at least their IDs are)
but might be disabled for some replicas, so to track their frontiers we
need yet another special case.