adapter: Add system compute instances
This commit adds the feature of system compute instances. Compute
instance IDs are now partitioned into two namespaces, User and System,
depending on what type of compute instance it is. We can now add
builtin compute instances and builtin compute instance replicas that
will be automatically added to the deployment on the next version
upgrade. The implementation is very similar to roles.

System compute instances will be used for system tasks such as
collecting billing information and serving simple common introspection
queries. Users will not be able to create anything in the system
compute instances or drop them.

Some parts of the role implementation was changed so that it would more
closely match the compute instance implementation.