storage: non-writing workers should not mess with shared upper in (storage) persist_sink
Before, there was a subtle bug in multi-worker ingestion pipelines: when
`environmentd` restarts it forces a re-creation of existing source
pipelines (potentially in an already-running `storaged`). The
`persist_sink` of the non-active write workers would update the write
frontier to the current upper of the shard, but then never change it
again (because they're not active). This led to frontier updates being
stuck, which led to queries being stuck.