persist: re-enable fast path for apply_merge_res in apply_diffs
Without this, we reconstruct the spine from scratch every time a
compaction result is merged. This is (1) wasteful, (2) may not match a
spine constructed incrementally and so may cause later merge responses
to not apply, and (3) prevents us from running a strict debug check that
a newly generated diff roundtrips.

Closes #14490.