persist: automatically heartbeat handles in background tasks (for now)
Ideally, users of persist read and write handles will do their own
explicit heartbeat-ing. In each place they've been used so far, there's
some nice place to hook a maybe_heartbeat call into.

However, we fixed a bunch of these in the release that just went out and
a set of new ones popped up in their place. The "correct" fix would be
to continue to track each of these down and address them individually,
but we've got enough fires as it is right now. Plus, until we figure out
a way to shake these out in CI, we'll continue to see them pop up in
production in new persist usages.

For now, introduce a background task for each read and each write handle
that automatically heartbeats it. The plan is to rip these back out at
some nebulous future point (probably not soon, to be honest), so
intentionally don't document this. Instead,  keep the other heartbeat
codepaths hooked up.

To make this more obvious, create ReadHandle::new and WriteHandle::new
methods (instead of creating them directly in lib.rs). We probably
should have done this a while ago, anyway.

Touches #15397