sql-parser: improve clarity of CREATE CONNECTION syntax
Add a strategic pair of parenthesis to CREATE CONNECTION to improve
clarity, separating the name of the connection type from the options.

Old:

  CREATE CONNECTION ... FOR KAFKA BROKER ..., PROGRESS TOPIC ...;

New:

  CREATE CONNECTION ... TO KAFKA (BROKER ..., PROGRESS TOPIC ...);

This will make it easier to visually distinguish connection types like
AWS vs AWS PRIVATELINK SERVICE.

The old syntax is still supported, in keeping with our backwards
compatibility guarantee. Documentation is updated, as are all tests that
are not explicitly testing the upgrade path from the old syntax. The
goal is to make it as hard as possible to find out about the old syntax.

We're early enough that we might yet be able to transition all of our
users to the new syntax, and eventually remove the old syntax entirely,
even though it is technically a breaking change.