stash: use inner peek result in upsert_key
upsert_key performed a peek_key_one then did some logic based on that. A
caller of upsert_key also performed the same peek, duplicating work.
Have upsert pass its found value to a Fn that can produce a new value
based on it.

Caveat: due to the memory stash, the initial peek result was is often
already cached, so the practical effect of this commit might be nil.