storage: use async OperatorBuilder for resumption operator
The previous implementation was using the old timely integration where
yielding without having seen any inputs leads to an busy loop of
operator activations.

I don't know if we had noticed in production but the resumption operator
was sleeping on an interval and had no inputs.

Additionally, the missed tick behaviour was set to the default which is
to burst when things are missed. This maybe have caused spikes in
persist calls if the dataflow was under load (e.g during an initial
snapshot) and then had the timer fire rapidly as many times as were
skipped.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>