storage: ensure reclock state moves in fully formed batches
We were previously pushing remap trace updates separately from pushing
upper frontier updates which means that we could end up in a situation
where reclock requests were being processed with partial state.

It's unclear if this could ever lead to wrong results but it's safer to
ensure that reclocking is only computed once the bindings of a given
time period have settled.

This PR also includes the following changes:

* The remap operator no longer ties its output frontier to the
  reclocked global source upper and it emits its binding batches as
  soon as they become availalbe. It is the responsibility of
  downstream operators to maintain an appropriate capability set
  depepending on the source frontier of *their* inputs.
* The reclock operator has been changed to use its initial capability
  to emit data which is only downgraded when its input source frontier
  downgrades. This should always succeeds or there is a bug, since the
  calculated source upper frontier must always advance.
* We were previously ignoring potential upper updates of the remap
  operator when calling `ReclockOperator::advance`. This is now fixed
  and uppers are also downgraded when we perform this operation.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>