stash-debug: add upgrade check
Teach the stash how to create and operate in a mode that allows
temporary writes that will never be committed to disk. This works by
using a single long transaction that uses SAVEPOINTs internally to nest
smaller transactions. The single outer transaction is never COMMIT'd,
thus when it disconnects all of the write intents will be removed. The
savepoint stash uses a low priority transaction to prevent a running
read-write stash from being blocked.

Add a subcommand to stash-debug that opens a catalog with migrations
using the savepoint stash.

Fixes #14681