[sinks] Create healthchecker and surface indefinite errors (#15586)
Modeled off of the source healthchecker, create a new sink healthchecker that reports indefinite errors to a persistent collection. At the moment, sinks only produce indefinite errors so this is relatively straighforward! We take the error, give it a timestamp at the current time (as determined by the storage controller) and then write it out to a persist shard.

A couple design decisions to note:

`@bkirwi` and I have talked about removing the initial snapshotting from the healthchecker. lmk if you'd like me to roll that into this change
I used an Arc<Mutex<Healthchecker>> within the kafka sink instead of something like Rc<RefCell<Healthchecker>>. While we don't strictly need the Arc<Mutex> (since the healthchecker is only used on a single thread), it was tricky to get both the borrow checker and clippy to allow compilation at the same time. There shouldn't be any lock contention and the code is more future-proofed since we don't have to worry about the single-thread restriction.
It's becoming increasingly clear that we'll want to reorganize the KafkaSinkState struct at some point -- there are a decent number of gymnastics to appease the borrow checker because different subfields have mutable borrows at overlapping points. All that said, I don't think rolling into this PR is the place to do this.
Motivation
#15163
#15164

Tips for reviewer
I've currently disabled the use of the introspection collection (as the source's is also currently disabled) but I did do a test run with it enabled: https://buildkite.com/materialize/tests/builds/43940