interchange: refactor combine_at_timestamp
The bug fixed in the last commit was the result of failing to keep
updates sorted by key... but the updates come out of the batch sorted by
key. This commit refactors combine_at_timestamp to operate on a
key-by-key basis, rather than mixing updates from all the keys together.

The refactor also partitions retractions and insertions into separate
vectors, to avoid needing to binary search to find the the point bewteen
the last retraction and the first insertion.