persist: increase write lease duration from 15m to 1h
Unlike read leases, which hold back since and seqno, there isn't a real
downside to longer write leases. It will only delay reclamation of
leaked blobs (by the leaked blob task we still haven't had a chance to
get to yet).

On the other hand, a longer write lease should help with periods of
transient network unavailability or high CPU usage (sufficiently high
that the heartbeat task doesn't get a chance to run).

It's become clear that we need more graceful handling of writer lease
timeouts, but in the meantime this should be a straightforward, quick
win to reduce the number of "WriterId ... was expired due to inactivity"
panics we see in prod.

Also halve the duration between heartbeats in the background task,
because we're making the lease duration so much longer.

Touches #15209