sql: fix database scoping in pg_catalog and information_schema
The views in pg_catalog and information_schema need to filter down the
information in mz_catalog relations to reflect just the current
database. We intended to accomplish this in #11292, but did so
incorrectly. The modification in #11292 added the join

    JOIN mz_catalog.mz_databases d ON (
        d.id IS NULL or d.name = pg_catalog.current_database()
    )

to every pg_catalog view, but this join is missing the condition that
relates mz_databases to the LHS of the join. The effect is that the
pg_catalog tables either return information about all databases, if the
`database` session variable names a valid database, or an empty set, if
the `database` session variable is invalid.

This commit fixes the bug and adds the tests that would have made the
bug evident. The pg_catalog views are still rather undertested, but we
have more pressing issues to fix.