persist: cache len on SpineBatch::Fueled
This gets called in Spine::tidy_layers, which gets called in of
Spine::introduce_batch. When we rehydrate a Spine from persist state,
introduce_batch is called for each of the HollowBatches on the state.
Combined, this resulted in an unintended and unnecessary repeated
re-computation of this len. (I haven't thought through exactly, but it's
possible this was `O(num_batches^2)` or something equally bad.)

This was significant enough to show up as the major factor during a high
CPU event on prod. It's easy enough to cache this len, so do it.