sql: don't add quotes when converting idents to values
We permit using identifiers as option values, e.g. `"xsmall"` instead of
`'xsmall'`. But we were accidentally including the quotes during
planning when the identifier could not be printed bare, by using
`to_string` instead of `into_string`.

The API here is just *asking* for this kind of confusion (`to_string` vs
`into_string` are mightly similar...), but fixing that is a larger
refactor than I want to tackle right now.